# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Vibe Coder Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /gigs:
    get:
      operationId: listGigs
      tags: [gigs]
      summary: List all gigs
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [task, hourly, build]
        - name: tag
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [open, closed, in_progress]
      responses:
        "200":
          description: List of gigs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Gig"
    post:
      operationId: createGig
      tags: [gigs]
      summary: Create a new gig
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGigBody"
      responses:
        "201":
          description: Created gig
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Gig"
  /gigs/{id}:
    get:
      operationId: getGig
      tags: [gigs]
      summary: Get a single gig
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Gig details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Gig"
        "404":
          description: Not found
    put:
      operationId: updateGig
      tags: [gigs]
      summary: Update a gig
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGigBody"
      responses:
        "200":
          description: Updated gig
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Gig"
    delete:
      operationId: deleteGig
      tags: [gigs]
      summary: Delete a gig
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── GIG REPLIES ─────────────────────────────────────────────────────────────
  /gigs/{id}/replies:
    get:
      operationId: listGigReplies
      tags: [gig-replies]
      summary: List replies for a gig
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of replies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GigReply"
    post:
      operationId: createGigReply
      tags: [gig-replies]
      summary: Submit a reply to a gig (no auth required)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGigReplyBody"
      responses:
        "201":
          description: Created reply
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GigReply"
  /gig-replies/{id}:
    delete:
      operationId: deleteGigReply
      tags: [gig-replies]
      summary: Delete a gig reply
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── FREELANCERS ─────────────────────────────────────────────────────────────
...
  /showcase:
    get:
      operationId: listShowcaseProjects
      tags: [showcase]
      summary: List all showcase projects
      parameters:
        - name: tag
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of showcase projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ShowcaseProject"
    post:
      operationId: createShowcaseProject
      tags: [showcase]
      summary: Submit a Vibe Coded project to the showcase
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateShowcaseProjectBody"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShowcaseProject"
  /showcase/{id}:
    get:
      operationId: getShowcaseProject
      tags: [showcase]
      summary: Get a showcase project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Showcase project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShowcaseProject"
        "404":
          description: Not found
    put:
      operationId: updateShowcaseProject
      tags: [showcase]
      summary: Update a showcase project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateShowcaseProjectBody"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShowcaseProject"
    delete:
      operationId: deleteShowcaseProject
      tags: [showcase]
      summary: Delete a showcase project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /showcase/{id}/upvote:
    post:
      operationId: upvoteShowcaseProject
      tags: [showcase]
      summary: Upvote a showcase project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated project with new upvote count
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShowcaseProject"

  # ─── TAGS ─────────────────────────────────────────────────────────────────────
  /tags:
    get:
      operationId: listTags
      tags: [tags]
      summary: List all tags
      responses:
        "200":
          description: List of tags
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tag"
    post:
      operationId: createTag
      tags: [tags]
      summary: Create a new tag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTagBody"
      responses:
        "201":
          description: Created tag
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tag"
  /tags/{id}:
    delete:
      operationId: deleteTag
      tags: [tags]
      summary: Delete a tag
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── STATS ───────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Gig:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [task, hourly, build]
        status:
          type: string
          enum: [open, closed, in_progress]
        tags:
          type: array
          items:
            type: string
        requirements:
          type: string
          nullable: true
        budgetMin:
          type: number
          nullable: true
        budgetMax:
          type: number
          nullable: true
        hourlyRate:
          type: number
          nullable: true
        recordingPath:
          type: string
          nullable: true
        publicSlug:
          type: string
        replyCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - description
        - type
        - status
        - tags
        - publicSlug
        - replyCount
        - createdAt
        - updatedAt
    CreateGigBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [task, hourly, build]
        tags:
          type: array
          items:
            type: string
        requirements:
          type: string
          nullable: true
        budgetMin:
          type: number
          nullable: true
        budgetMax:
          type: number
          nullable: true
        hourlyRate:
          type: number
          nullable: true
        recordingPath:
          type: string
          nullable: true
      required:
        - title
        - description
        - type
        - tags
    UpdateGigBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [task, hourly, build]
        status:
          type: string
          enum: [open, closed, in_progress]
        tags:
          type: array
          items:
            type: string
        requirements:
          type: string
          nullable: true
        budgetMin:
          type: number
          nullable: true
        budgetMax:
          type: number
          nullable: true
        hourlyRate:
          type: number
          nullable: true
        recordingPath:
          type: string
          nullable: true

    # ─── GIG REPLY ────────────────────────────────────────────────────────────
    GigReply:
      type: object
      properties:
        id:
          type: integer
        gigId:
          type: integer
        senderName:
          type: string
        message:
          type: string
          nullable: true
        voiceNotePath:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - gigId
        - senderName
        - createdAt
    CreateGigReplyBody:
      type: object
      properties:
        senderName:
          type: string
        message:
          type: string
          nullable: true
        voiceNotePath:
          type: string
          nullable: true
      required:
        - senderName

    # ─── FREELANCER ───────────────────────────────────────────────────────────
...
    ShowcaseProject:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        url:
          type: string
          nullable: true
        screenshotPath:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        tools:
          type: array
          items:
            type: string
        builderName:
          type: string
        upvotes:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - description
        - tags
        - tools
        - builderName
        - upvotes
        - createdAt
        - updatedAt
    CreateShowcaseProjectBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        url:
          type: string
          nullable: true
        screenshotPath:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        tools:
          type: array
          items:
            type: string
        builderName:
          type: string
      required:
        - name
        - description
        - tags
        - tools
        - builderName
    UpdateShowcaseProjectBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        url:
          type: string
          nullable: true
        screenshotPath:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        tools:
          type: array
          items:
            type: string
        builderName:
          type: string

    # ─── TAGS ─────────────────────────────────────────────────────────────────
    Tag:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
          enum: [tool, skill, gig_type, general]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - category
        - createdAt
    CreateTagBody:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
          enum: [tool, skill, gig_type, general]
      required:
        - name
        - category

    # ─── STATS ────────────────────────────────────────────────────────────────
...